* feat/factory_fw/doc_virgil:
  doc:[factory_fw] update doc.
  style:[factory_fw] Move “TF_DATA_TYPE_DUALIMAGE_WITH_INFERENCE” data definition; change the type field to uint32_t.
  doc: [factory_fw] update the doc for uart alarm
  doc:[factory_fw] update doc.
  style：Modify variable name, Add some comments.
  doc:[factory_fw] add architecture_CN.md

+++ examples/factory_firmware/main/task_flow_module/tf_module_uart_alarm.c
@@ -24,7 +24,7 @@ static void __event_handler(void *handler_args, esp_event_base_t base, int32_t i
 {
     tf_module_uart_alarm_t *p_module_ins = (tf_module_uart_alarm_t *)handler_args;
    
-    uint8_t type = ((uint8_t *)p_event_data)[0];
+    uint32_t type = ((uint32_t *)p_event_data)[0];
     if( type !=  TF_DATA_TYPE_DUALIMAGE_WITH_AUDIO_TEXT) {
         ESP_LOGW(TAG, "unsupported type %d", type);
         tf_data_free(p_event_data);
